Version 6.2.7.
Fixes to long-standing assignment problem.  Previously, the command
    set var X = EXPR
would work only for EXPRs that did not involve restarting the thread.
If EXPR contained a method call or the 'new' operator, in particular,
the assignment would fail, because evaluation of the local variable
X into an lvalue would save the frame at the time of that evaluation
in the LValueLocal object representing the variable. Then EXPR would 
be evaluated, invalidating the frame, and only then did the assignment
take place.  The fix is to save the frame-getter (type GetFrame) instead
of the frame.

Bring NEWS up-to-date a bit.



git-svn-id: file:///home/cs/hilfingr/SVN/gjdb/trunk@1353 8fc76cf4-f431-0410-aa1a-800000000000